First calculate function added
This first calculate function is not finished, but it does contain the largest part of the calculation. The calculation checks whether freespace overlaps with coordinates.

The extra freespace has yet to be added and removed (if wrong) from the plot. In addition, it must be checked whether freespaces extend beyond the range of the spacelist (for example negative numbers or numbers greater than 180 in width and higher than 160 in height).