Calculate function, not finished
This calculate function is not finished, but it does contain the largest part of the calculation. The calculation checks whether freespace overlaps with coordinates.

Sometimes the calculate function checks good, but the calculate function does not always end up in the 'within' statement. As a result, the function comes in an infinite loop and continues to add infinite freespace. My suspicion is that this happens as soon as numbers come in the minus. The counter (which counts per meter of extra freespace) does not check for bounderies yet. The 'hello' and 'hoi' prints are there to better understand what this function does.